Added I/O generation and testing
With 10,000 iterations on randomized input, the network correctly leans
towards the correct result (argmax of the output vector matches the
expected output 100% - tried it with 2 random seeds), but there's
a fairly big gap between the expected activation vs. actual activation
of the neuron representing the addition answer.

Another problem in the way the input is currently being fed is that the
input spans the entire problem space. So, it's difficult to claim that
the network has 'understood addition' rather than just memomrizing
results. Suggested improvement:
* Train the network on a subset of possible inputs
* Test it on inputs it hasn't seen